+++ src/main/java/site/imcu/tape/security/jwt/TokenProvider.java
@@ -95,15 +95,12 @@ public class TokenProvider implements InitializingBean {
 
    public site.imcu.tape.pojo.User getCurrentUser(){
       Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-//      Object principal = authentication.getPrincipal();
-//      site.imcu.tape.pojo.User user1 = new site.imcu.tape.pojo.User();
-//      BeanUtils.copyProperties(principal,user1);
       Object credentials = authentication.getCredentials();
       Claims claims = Jwts.parser()
               .setSigningKey(key)
               .parseClaimsJws(credentials.toString())
               .getBody();
-      Integer userId = claims.get("userId", Integer.class);
+      Long userId = claims.get("userId", Long.class);
       String username = claims.get("username", String.class);
       site.imcu.tape.pojo.User user = new site.imcu.tape.pojo.User();
       user.setId(userId);
